fix(mypy): Complete mypy type coverage - achieve zero errors (92 → 0)
Major fixes in final batches:

Batch 8 (5 errors fixed - SQLAlchemy Base subclass):
- Added type: ignore[misc] to 5 SQLAlchemy ORM model classes
- statistics.py, settings.py, schema.py, fingerprint.py (2 classes)
- Issue: declarative_base() returns Any type in mypy, causing inheritance errors

Batch 9 (4 errors fixed - Quality assessor overrides):
- Added type: ignore[override] to 4 method signatures
- dynamic_assessment.py: assess() and detailed_analysis()
- distortion_assessment.py: assess() and detailed_analysis()
- Specialized assessors intentionally have different signatures than base class

Batch 10 (All remaining 7 errors fixed):
- preset_parameters.py:221 - Added str() conversion in dict comprehension
- proactive_buffer.py:77 - Added Tuple[Path, np.ndarray] type parameter to cast()
- feature_extractor.py:183 - Added float() conversions for np.std() and np.mean()
- audio_fingerprint_analyzer.py:124 - Added Dict[str, Any] annotation + Any import
- fingerprint_repository.py:566 - Changed return from fingerprint_dict to None
- startup.py:114 - Added return-value to type: ignore[arg-type, return-value]
- fingerprint_extractor.py:168 - Kept type: ignore[import-untyped] for requests

Session Summary:
- Started: 92 mypy errors (from previous session)
- Ended: 0 mypy errors
- Total reduction in session: 92 errors → 0 (100% complete!)
- Total reduction from start of type audit: 440 → 0 (100%)

Pattern-based fixes applied:
1. NumPy type conversions (float(), int())
2. SQLAlchemy ORM type suppression
3. Method override signature matching
4. Generic type parameter specifications
5. Dict/tuple type annotations

+++ auralis/analysis/ml/feature_extractor.py
@@ -178,8 +178,8 @@ class FeatureExtractor:
         if len(tempos) < 2:
             return 0.5
 
-        tempo_std = np.std(tempos)
-        tempo_mean = np.mean(tempos)
+        tempo_std = float(np.std(tempos))
+        tempo_mean = float(np.mean(tempos))
         return 1.0 - min(tempo_std / (tempo_mean + 1e-10), 1.0)
 
     def _onset_rate(self, audio: np.ndarray) -> float:
